Route Args Resolver Middleware
Resolve Tournament, Category, Area, Participant from route arguments in a dedicated middleware, so those entities are available via $request.
This way, they don't have to be resolved in every controller.

Introduce new Exception "EntityNotFoundException" that can be thrown whenever the entity addressed in the url parameters does not exist.
It will be caught in the ErrorHandlingMiddleware and display a dedicated "not found" page.